[Decoder] Refactore content of decoder.rs
